requirements.txt
asgiref==3.7.2
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
joblib==1.3.0
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.6
py4j==0.10.9.7
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.2.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
